+++ src/main/java/com/infoshare/alpha/wwr/domain/patients/readmodel/PatientsReadModelDb.java
@@ -1,4 +1,4 @@
-//package com.infoshare.alpha.wwr.domain.patients.readmodel;
+//package com.infoshare.alpha.wwr.domain.patients.dao;
 //
 //import com.infoshare.alpha.wwr.domain.patients.entity.Patients;
 //import com.infoshare.alpha.wwr.domain.patients.query.PatientQuery;
