+++ src/main/java/com/infoshare/alpha/wwr/utils/PatientsFacilitiesController.java
@@ -1,7 +1,7 @@
 package com.infoshare.alpha.wwr.utils;
 
-import com.infoshare.alpha.wwr.common.Facility;
-import com.infoshare.alpha.wwr.domain.facilities.FacilitiesService;
+import com.infoshare.alpha.wwr.domain.facilities.entity.Facility;
+import com.infoshare.alpha.wwr.domain.facilities.service.FacilitiesService;
 import com.infoshare.alpha.wwr.domain.facilities.query.FacilityPatientQuery;
 import com.infoshare.alpha.wwr.domain.facilities.query.FacilityQueryField;
 import com.infoshare.alpha.wwr.domain.patients.entity.Patient;
